Reorganize alchemist/compound
