+++ lib/alchemist/compound.rb
@@ -1,7 +1,8 @@
-# TODO: remove all of this
-Alchemist.register_operation_conversions(:distance, :distance, :*, :square_meters)
-Alchemist.register_operation_conversions(:area, :distance, :*, :cubic_meters)
-Alchemist.register_operation_conversions(:distance, :area, :*, :cubic_meters)
-Alchemist.register_operation_conversions(:area, :distance, :/, :meters)
-Alchemist.register_operation_conversions(:volume, :distance, :/, :square_meters)
-Alchemist.register_operation_conversions(:volume, :area, :/, :meters)
+module Alchemist
+  register_operation_conversions(:distance, :distance, :*, :square_meters)
+  register_operation_conversions(:area, :distance, :*, :cubic_meters)
+  register_operation_conversions(:distance, :area, :*, :cubic_meters)
+  register_operation_conversions(:area, :distance, :/, :meters)
+  register_operation_conversions(:volume, :distance, :/, :square_meters)
+  register_operation_conversions(:volume, :area, :/, :meters)
+end
